requirements.txt
elevenlabs[pyaudio]
fastapi
uvicorn
python-dotenv
websockets
numpy
sounddevice
python-multipart
